moving KeyHasher from platform.sanbox.ledger.sql to daml.lf.value + rename it ValueHasher
